cloud_functions/parse_ibjjf_results_webpage/requirements.txt
bs4
google-cloud-storage==1.41.0
